Refactor `BuildServiceProvider<T, P>` into an interface
By introducing the `DefaultBuildServiceProvider<T, P>` implementation class.
